requirements.txt
appnope==0.1.2
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
boto3==1.21.12
botocore==1.24.12
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.3
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-storages==1.12.3
djangorestframework==3.13.1
executing==0.8.3
ipython==8.1.1
jedi==0.18.1
jmespath==0.10.0
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
urllib3==1.26.8
wcwidth==0.2.5
whitenoise==6.0.0
